Edit modal updated

+++ backend/src/index.js
@@ -2,6 +2,7 @@ import express from "express";
 import dotenv from "dotenv";
 import cors from "cors";
 import Item from "./models/item.model.js";
+import mongoose from "mongoose";
 // import itemRoutes from "./routes/items.route.js";
 
 import { connectDB } from "./lib/db.js";
@@ -27,14 +28,20 @@ app.get('/', (req, res) => {
 })
 
 app.get('/getItem/:id', (req, res) => {
-  const id = req.params.id
+  const id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ error: "Invalid ObjectId" });
+  }
   Item.findById(id)
-  .then(items => res.json(items))
-  .catch(err => console.log(err))
-})
+    .then(item => res.json(item))
+    .catch(err => res.status(500).json(err));
+});
 
 app.put('/updateItem/:id', (req, res) => {
   const id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ error: "Invalid ObjectId" });
+  }
   Item.findByIdAndUpdate(id, req.body, { new: true })
     .then(updatedItem => res.json(updatedItem))
     .catch(err => {
@@ -51,6 +58,9 @@ app.post("/createItem", (req, res) => {
 
 app.delete('/deleteItem/:id', (req, res) => {
   const id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ error: "Invalid ObjectId" });
+  }
   Item.findByIdAndDelete(id)
     .then(result => res.json(result))
     .catch(err => {
